Added models and repos

+++ src/main/java/com/example/springsecurity/repository/CustomerRepository.java
@@ -1,5 +1,6 @@
 package com.example.springsecurity.repository;
 
+
 import com.example.springsecurity.model.Customer;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
@@ -7,7 +8,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CustomerRepository extends CrudRepository<Customer,Integer> {
+public interface CustomerRepository extends CrudRepository<Customer, Long> {
 
     List<Customer> findByEmail(String email);
+    
 }
